refactor: replace intToString func to utility folder
